Use resin/rpi-raspbian as the base image
I want to run the SFTP image on a Pi. I don't know if the resin.io
image is a better choice than the armhf images or not, but I've
confirmed that I can build it and successfully login with
username/password.

Some of the tests are failing, but they're also failing on OS 10.11.6,
so this issue doesn't appear to be specific to the new ARM image.